Refactor and relocate FathomNet crop tools
Migrated FathomNet crop-related classes and utilities from `DD_tools` to `TreeOfLife_toolbox`, ensuring proper namespace adjustments and functionality preservation. Added a README to document the new module structure and enhanced docstrings for better clarity on cropping, scheduling, and filtering operations.
